feat(FX-4806): price control for create / edit alert flows
