+++ src/app/Components/PriceRange/types.ts
@@ -1,0 +1,3 @@
+import { Numeric } from "app/Components/ArtworkFilter/Filters/helpers"
+
+export type PriceRange = Numeric[]
