+++ src/app/Scenes/SavedSearchAlert/CreateSavedSearchAlert.tsx
@@ -48,7 +48,14 @@ export const CreateSavedSearchAlert: React.FC<CreateSavedSearchAlertProps> = (pr
                 initialParams={params}
               />
               <Stack.Screen name="EmailPreferences" component={EmailPreferencesScreen} />
-              <Stack.Screen name="AlertPriceRange" component={AlertPriceRangeScreen} />
+              <Stack.Screen
+                name="AlertPriceRange"
+                component={AlertPriceRangeScreen}
+                options={{
+                  // Avoid PanResponser conflicts between the slider and the slide back gesture
+                  gestureEnabled: false,
+                }}
+              />
             </Stack.Navigator>
           </Box>
         </FancyModal>
